• refactor(api): remove legacy http client and global state keys
