+++ src/queries/index.ts
@@ -1,5 +1,4 @@
 export * from './categories';
-export * from './globalState';
 export * from './messages';
 export * from './offers';
 export * from './requests';
