--- src/services/api/index.ts
@@ -1,6 +1,0 @@
-import appFetch, { BASE_URL } from './httpClient';
-
-export type { AppFetchOptions } from './httpClient';
-
-export { BASE_URL, appFetch };
-export default appFetch;
