+++ src/services/auth.service.js
@@ -1,4 +1,4 @@
-import appFetch from './api';
+import appFetch from '../utilities/appFetch';
 
 const login = async (username, password) => {
   const { auth_hash } = await appFetch('/auth/', {
